refactor: renmae borg folder to app
