+++ backend/app/worker/prune.go
@@ -1,8 +1,8 @@
 package worker
 
 import (
-	"arco/backend/borg/types"
-	"arco/backend/borg/util"
+	"arco/backend/app/types"
+	"arco/backend/app/util"
 	"context"
 	"fmt"
 	"os/exec"
