+++ backend/app/worker/worker.go
@@ -1,8 +1,8 @@
 package worker
 
 import (
-	"arco/backend/borg/types"
-	"arco/backend/borg/util"
+	"arco/backend/app/types"
+	"arco/backend/app/util"
 	"context"
 	"go.uber.org/zap"
 )
